refactor(platform): extract api-gateway route policy into pkg/platform/routepolicy (#830)
Second seam of the Platform decomposition (#756), after the auth-identity
extraction (#828). Move the api-gateway route-authorization adapter off the
Platform god-object into a domain-named package with an explicit-deps
constructor and direct unit tests.

- pkg/platform/routepolicy: New(Deps) *Policy builds the policy that resolves a
  caller's roles and checks (connection, method, path) against the persona
  authorizer's APIRoutes rules. New returns the concrete *Policy (accept
  interfaces, return structs); the package imports only middleware and persona,
  deliberately NOT pkg/toolkits/apigateway — the SetRoutePolicy call site in
  platform.go is the compile-time proof that *Policy satisfies
  apigatewaykit.RoutePolicy, so the package does not depend on the toolkit it
  serves.
- platform.go: apiGatewayRoutePolicy and its Allow/resolveRoles are removed;
  WireAPIGatewayRoutePolicy delegates to routepolicy.New with its guards (nil
  authorizer, non-persona authorizer) and registry-walk injection unchanged.
- The pre-existing direct adapter test moves into routepolicy (the resolveRoles
  preference order, the panic-authenticator no-double-verify regression) and
  gains the Authenticator-fallback case the original omitted — routepolicy is
  100% covered. The WireAPIGatewayRoutePolicy integration tests still pass.
- Import ratchet regenerated for the new platform -> routepolicy and
  routepolicy -> middleware/persona edges.

Closes #830. Part of #756.
