+++ pkg/platform/platform_test.go
@@ -5054,137 +5054,3 @@ func TestWireAPIGatewayRoutePolicy_NoAuthorizer_NoOp(t *testing.T) {
 	p.authorizer = nil // simulate "no authorizer"
 	p.WireAPIGatewayRoutePolicy()
 }
-
-// panicAuthenticator fails the test if its Authenticate is called.
-// Used to prove the apiGatewayRoutePolicy doesn't redundantly call
-// Authenticate when PlatformContext.Roles is already populated by
-// the upstream MCP middleware (the common hot-path case).
-type panicAuthenticator struct{}
-
-func (panicAuthenticator) Authenticate(_ context.Context) (*middleware.UserInfo, error) {
-	panic("Authenticate called when PlatformContext.Roles should have been used")
-}
-
-// TestAPIGatewayRoutePolicy_Allow_DirectCases exercises the closure
-// that bridges persona.Authorizer onto apigateway.RoutePolicy. Three
-// scenarios: pre-auth user with allow rule passes, pre-auth user with
-// deny rule is refused, no roles in ctx hits the fail-closed path.
-func TestAPIGatewayRoutePolicy_Allow_DirectCases(t *testing.T) {
-	personaReg := persona.NewRegistry()
-	if err := personaReg.Register(&persona.Persona{
-		Name:  "analyst",
-		Roles: []string{"analyst"},
-		Tools: persona.ToolRules{Allow: []string{"*"}},
-		APIRoutes: []persona.APIRouteRule{
-			{Connection: "crm", Methods: []string{"GET"}, Paths: []string{"/v1/users/*"}},
-			{Connection: "crm", Methods: []string{"DELETE"}, Action: persona.ActionDeny},
-		},
-	}); err != nil {
-		t.Fatalf("register persona: %v", err)
-	}
-	mapper := &persona.OIDCRoleMapper{Registry: personaReg}
-	authzr := persona.NewAuthorizer(personaReg, mapper)
-	policy := apiGatewayRoutePolicy{authn: nil, pa: authzr}
-
-	t.Run("allow rule matches → true", func(t *testing.T) {
-		ctx := middleware.WithPreAuthenticatedUser(context.Background(), &middleware.UserInfo{
-			UserID: "u1", Roles: []string{"analyst"},
-		})
-		allowed, reason := policy.Allow(ctx, "crm", "GET", "/v1/users/123")
-		if !allowed {
-			t.Errorf("got allowed=false, reason=%q; want allowed=true", reason)
-		}
-	})
-
-	t.Run("deny rule matches → false with reason", func(t *testing.T) {
-		ctx := middleware.WithPreAuthenticatedUser(context.Background(), &middleware.UserInfo{
-			UserID: "u1", Roles: []string{"analyst"},
-		})
-		allowed, reason := policy.Allow(ctx, "crm", "DELETE", "/v1/users/123")
-		if allowed {
-			t.Error("DELETE was allowed despite deny rule")
-		}
-		if reason == "" {
-			t.Error("expected non-empty reason on denial")
-		}
-	})
-
-	t.Run("connection with no APIRoutes for resolved persona → passthrough", func(t *testing.T) {
-		// Empty roles resolve to DefaultPersona (no APIRoutes). The
-		// route policy is layered on top of the platform's main
-		// IsAuthorized gate; when the persona has no APIRoutes
-		// touching the connection the route check is a no-op (true).
-		// Fail-closed for unauthenticated callers is the platform's
-		// MCPToolCallMiddleware's job, not this adapter's.
-		allowed, _ := policy.Allow(context.Background(), "crm", "GET", "/v1/users/1")
-		if !allowed {
-			t.Error("expected passthrough for persona without APIRoutes for crm")
-		}
-	})
-
-	t.Run("PlatformContext roles preferred over Authenticator (no double-verify)", func(t *testing.T) {
-		// Confirms the resolveRoles preference: MCPToolCallMiddleware's
-		// PlatformContext.Roles is read instead of forcing a second
-		// Authenticate call. Use an Authenticator that would PANIC if
-		// invoked — if the test passes, the role-resolution path
-		// reused the pre-extracted roles from PlatformContext.
-		policyNoAuthn := apiGatewayRoutePolicy{
-			authn: panicAuthenticator{},
-			pa:    authzr,
-		}
-		pc := middleware.NewPlatformContext("req-1")
-		pc.UserID = "u1"
-		pc.Roles = []string{"analyst"}
-		ctx := middleware.WithPlatformContext(context.Background(), pc)
-		allowed, _ := policyNoAuthn.Allow(ctx, "crm", "GET", "/v1/users/123")
-		if !allowed {
-			t.Error("PlatformContext.Roles path did not produce allow decision")
-		}
-	})
-
-	t.Run("authenticated user with empty roles still uses PlatformContext (no double-verify)", func(_ *testing.T) {
-		// Regression for the resolveRoles guard: if the predicate is
-		// `len(pc.Roles) > 0`, an authenticated user with no role
-		// claims falls through to the Authenticator fallback —
-		// reintroducing the redundant signature verification this
-		// path was meant to avoid. The correct signal is "auth
-		// middleware ran" (pc.UserID populated), not "user has
-		// roles". panicAuthenticator catches the regression.
-		policyNoAuthn := apiGatewayRoutePolicy{
-			authn: panicAuthenticator{},
-			pa:    authzr,
-		}
-		pc := middleware.NewPlatformContext("req-2")
-		pc.UserID = "u2"
-		pc.Roles = nil // authenticated but role-less
-		ctx := middleware.WithPlatformContext(context.Background(), pc)
-		// We don't care about the Allow outcome — only that
-		// Authenticate is NOT called (panic would fail the test).
-		_, _ = policyNoAuthn.Allow(ctx, "crm", "GET", "/v1/users")
-	})
-
-	t.Run("connection-targeted persona but auth context absent → fail-closed", func(t *testing.T) {
-		// Build a custom registry where the default persona has
-		// APIRoutes touching crm — empty roles still resolve to it,
-		// so the route check actually applies, and an unmatched
-		// (method, path) is denied.
-		reg := persona.NewRegistry()
-		def := &persona.Persona{
-			Name:  "default",
-			Roles: []string{},
-			Tools: persona.ToolRules{Allow: []string{"*"}},
-			APIRoutes: []persona.APIRouteRule{
-				{Connection: "crm", Methods: []string{"GET"}, Paths: []string{"/v1/users/*"}},
-			},
-		}
-		_ = reg.Register(def)
-		reg.SetDefault("default")
-		mapper2 := &persona.OIDCRoleMapper{Registry: reg}
-		auth2 := persona.NewAuthorizer(reg, mapper2)
-		policy2 := apiGatewayRoutePolicy{authn: nil, pa: auth2}
-		allowed, _ := policy2.Allow(context.Background(), "crm", "DELETE", "/v1/anything")
-		if allowed {
-			t.Error("DELETE on path-restricted connection allowed without explicit allow rule")
-		}
-	})
-}
